+++ aiolancium/utilities/utilities.py
@@ -3,6 +3,29 @@ from hashlib import md5
 
 from typing import Iterable
 
+import os
+
+
+async def upload_helper(
+    awaitable_create_method,
+    awaitable_upload_method,
+    path,
+    source,
+    source_type,
+    chunk_size,
+    **kwargs,
+):
+    file_size = os.path.getsize(source)
+
+    await awaitable_create_method(
+        path=path, source=source, source_type=source_type, size=file_size, **kwargs
+    )
+
+    for chunk_data in read_binary_chunks_from_file(
+        file_name=source, chunk_size=chunk_size
+    ):
+        await awaitable_upload_method(path, **chunk_data)
+
 
 def extract_kwargs(keys: Iterable, kwargs):
     return {key: kwargs.pop(key) for key in keys if key in kwargs}